move key fetch logic to identity component
